+++ src/Core/Sampler.hpp
@@ -23,36 +23,36 @@ class Sampler {
   Sampler(const CConfig& config) : Sampler(config.imageHeight, config.imageWidth, config.samplesPerPixel) {}
 
   Point2 getRandomSample(ushort pixelX, ushort pixelY) const {
-    const float u = (pixelX + Random::fraction()) / (imageWidth - 1);
-    const float v = (pixelY + Random::fraction()) / (imageHeight - 1);
+    const double u = (pixelX + Random::fraction()) / (imageWidth - 1);
+    const double v = (pixelY + Random::fraction()) / (imageHeight - 1);
     return Point2(u, v);
   }
 
   // MLT
-  Point2 getRandomSample(ushort pixelX, ushort pixelY, float random1, float random2) const {
-    const float u = (pixelX + random1) / (imageWidth - 1);
-    const float v = (pixelY + random2) / (imageHeight - 1);
+  Point2 getRandomSample(ushort pixelX, ushort pixelY, double random1, double random2) const {
+    const double u = (pixelX + random1) / (imageWidth - 1);
+    const double v = (pixelY + random2) / (imageHeight - 1);
     return Point2(u, v);
   }
 
   Point2 getUniformSample(ushort pixelX, ushort pixelY, ushort index) const {
     const ushort i = index % samplesPerEdge;
     const ushort j = index / samplesPerEdge;
-    const float offsetX = i / samplesPerEdge + 1 / (2 * samplesPerEdge);
-    const float offsetY = j / samplesPerEdge + 1 / (2 * samplesPerEdge);
+    const double offsetX = i / samplesPerEdge + 1 / (2 * samplesPerEdge);
+    const double offsetY = j / samplesPerEdge + 1 / (2 * samplesPerEdge);
 
-    const float u = (pixelX + offsetX) / (imageWidth - 1);
-    const float v = (pixelY + offsetY) / (imageHeight - 1);
+    const double u = (pixelX + offsetX) / (imageWidth - 1);
+    const double v = (pixelY + offsetY) / (imageHeight - 1);
     return Point2(u, v);
   }
 
   Point2 getStratifiedSample(ushort pixelX, ushort pixelY, ushort index) const {
     const ushort i = index % samplesPerEdge;
     const ushort j = index / samplesPerEdge;
-    const float offsetX = i / samplesPerEdge + 1 / (2 * samplesPerEdge) + (Random::fraction() - 0.5) / samplesPerEdge;
-    const float offsetY = j / samplesPerEdge + 1 / (2 * samplesPerEdge) + (Random::fraction() - 0.5) / samplesPerEdge;
-    const float u = (pixelX + offsetX) / (imageWidth - 1);
-    const float v = (pixelY + offsetY) / (imageHeight - 1);
+    const double offsetX = i / samplesPerEdge + 1 / (2 * samplesPerEdge) + (Random::fraction() - 0.5) / samplesPerEdge;
+    const double offsetY = j / samplesPerEdge + 1 / (2 * samplesPerEdge) + (Random::fraction() - 0.5) / samplesPerEdge;
+    const double u = (pixelX + offsetX) / (imageWidth - 1);
+    const double v = (pixelY + offsetY) / (imageHeight - 1);
     return Point2(u, v);
   }
 };
