+++ src/GeoObjects/GeometricalObject.hpp
@@ -7,11 +7,11 @@
 
 class GeometricalObject {
  public:
-  virtual bool intersect(const Ray& ray, float tMin, float tMax, SInteraction& interaction) const = 0;
-  virtual bool computeBoundingBox(float t0, float t1, AABB& outputBox) const = 0;
+  virtual bool intersect(const Ray& ray, double tMin, double tMax, SInteraction& interaction) const = 0;
+  virtual bool computeBoundingBox(double t0, double t1, AABB& outputBox) const = 0;
   virtual Point3 samplePoint() const { return Point3(0, 0, 0); }
-  virtual Point3 samplePoint(float random1, float random2) const { return Point3(0, 0, 0); }
-  virtual float getArea() const { return 0; }
+  virtual Point3 samplePoint(double random1, double random2) const { return Point3(0, 0, 0); }
+  virtual double getArea() const { return 0; }
   virtual std::shared_ptr<Material> getMaterial() const { return nullptr; }
 };
