+++ src/Materials/Metal.hpp
@@ -10,13 +10,13 @@
 class Metal : public Material {
  private:
   Color albedo;
-  float fuzz;
+  double fuzz;
 
   // Dummy variable
-  float glossiness = 25.0;
+  double glossiness = 25.0;
 
  public:
-  Metal(const Color& albedo, float fuzz) : albedo(albedo), fuzz(Math::clamp(fuzz, 0.0, 1.0)) {}
+  Metal(const Color& albedo, double fuzz) : albedo(albedo), fuzz(Math::clamp(fuzz, 0.0, 1.0)) {}
 
   virtual bool scatter(const Ray& incoming, const SInteraction& interaction, Color& attenuation,
                        Ray& scattered) const override {
@@ -27,15 +27,15 @@ class Metal : public Material {
     return (dot(scattered.direction, interaction.normal) > 0);
   }
 
-  float brdf(const Vector3& wi, const Vector3& n, const Vector3& wo) const override {
-    const float won = dot(wo, n);
-    const float win = dot(wi, n);
+  double brdf(const Vector3& wi, const Vector3& n, const Vector3& wo) const override {
+    const double won = dot(wo, n);
+    const double win = dot(wi, n);
     const Vector3 reflected = (-wi).reflect(n);
     return (glossiness + 2.0) / (2.0 * Math::pi) * pow(std::fmax(dot(reflected, wo), 0.0), glossiness) /
            std::fmax(std::abs(win), std::abs(won));
   }
 
-  float pdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override {
+  double pdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override {
     const Vector3 reflected = (-wi).reflect(normal);
     return (glossiness + 1.0) / (2.0 * Math::pi) * std::pow(std::fmax(dot(reflected, wo), 0.0), glossiness);
   }
